Updated the daemon to watch the file system
Right now it watches the root directory instead of important
subdirectories.

TODO: fork off a thread to watch the outbox subdirectory. Change the
init/update functions to collect messages from the outbox and send
them. (probably init/update can be treated the same; the only difference
is whether the file was created before the daemon was launched)